+++ zebra/if_ioctl.c
@@ -29,7 +29,6 @@
 #include "ioctl.h"
 #include "connected.h"
 #include "memory.h"
-#include "zebra_memory.h"
 #include "log.h"
 #include "vrf.h"
 #include "vty.h"
