+++ zebra/kernel_socket.c
@@ -32,7 +32,6 @@
 #include "sockunion.h"
 #include "connected.h"
 #include "memory.h"
-#include "zebra_memory.h"
 #include "ioctl.h"
 #include "log.h"
 #include "table.h"
