+++ zebra/main.c
@@ -26,7 +26,6 @@
 #include "thread.h"
 #include "filter.h"
 #include "memory.h"
-#include "zebra_memory.h"
 #include "prefix.h"
 #include "log.h"
 #include "plist.h"
