+++ zebra/rtadv.c
@@ -23,7 +23,6 @@
 #include <zebra.h>
 
 #include "memory.h"
-#include "zebra_memory.h"
 #include "sockopt.h"
 #include "thread.h"
 #include "if.h"
