+++ zebra/rtread_sysctl.c
@@ -24,7 +24,6 @@
 #if !defined(GNU_LINUX)
 
 #include "memory.h"
-#include "zebra_memory.h"
 #include "log.h"
 #include "vrf.h"
