+++ zebra/zebra_ptm_redistribute.c
@@ -26,7 +26,6 @@
 #include "zebra/zapi_msg.h"
 #include "zebra/zebra_ptm.h"
 #include "zebra/zebra_ptm_redistribute.h"
-#include "zebra/zebra_memory.h"
 
 static int zsend_interface_bfd_update(int cmd, struct zserv *client,
 				      struct interface *ifp, struct prefix *dp,
