+++ zebra/zebra_routemap.c
@@ -21,7 +21,6 @@
 #include <zebra.h>
 
 #include "memory.h"
-#include "zebra_memory.h"
 #include "prefix.h"
 #include "rib.h"
 #include "vty.h"
