+++ zebra/zebra_router.c
@@ -25,7 +25,6 @@
 #include "lib/frratomic.h"
 
 #include "zebra_router.h"
-#include "zebra_memory.h"
 #include "zebra_pbr.h"
 #include "zebra_vxlan.h"
 #include "zebra_mlag.h"
@@ -33,6 +32,7 @@
 #include "debug.h"
 
 DEFINE_MTYPE_STATIC(ZEBRA, RIB_TABLE_INFO, "RIB table info");
+DEFINE_MTYPE_STATIC(ZEBRA, ZEBRA_RT_TABLE, "Zebra VRF table");
 
 struct zebra_router zrouter = {
 	.multipath_num = MULTIPATH_NUM,
@@ -121,7 +121,7 @@ struct route_table *zebra_router_get_table(struct zebra_vrf *zvrf,
 	if (zrt)
 		return zrt->table;
 
-	zrt = XCALLOC(MTYPE_ZEBRA_NS, sizeof(*zrt));
+	zrt = XCALLOC(MTYPE_ZEBRA_RT_TABLE, sizeof(*zrt));
 	zrt->tableid = tableid;
 	zrt->afi = afi;
 	zrt->safi = safi;
@@ -185,7 +185,7 @@ static void zebra_router_free_table(struct zebra_router_table *zrt)
 	RB_REMOVE(zebra_router_table_head, &zrouter.tables, zrt);
 
 	XFREE(MTYPE_RIB_TABLE_INFO, table_info);
-	XFREE(MTYPE_ZEBRA_NS, zrt);
+	XFREE(MTYPE_ZEBRA_RT_TABLE, zrt);
 }
 
 void zebra_router_release_table(struct zebra_vrf *zvrf, uint32_t tableid,
